requirements.txt
pandas
yfinance
sqlalchemy
psycopg2-binary
apache-airflow
matplotlib